requirements.txt
behave==1.2.6
PyHamcrest==2.0.2
python-dateutil==2.6.1
robotframework==3.2.2
selenium==3.141.0
virtualenv==16.7.5
webdrivermanager==0.9.0
